+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   continue to be declared in a single *decl* field as before.  This
   makes it easier to define opaque structs where the members are not
   listed. Useful with *wrap_struct_as=class*.
+- Support `base: struct` in the `typemap` field of the YAML file. This allows structs wrapped
+  outside of the current YAML file to be used.
 
 ### Fixed
 - Order of header files in *cxx_header* is preserved in the generated code.